third_party/WebKit/LayoutTests/platform/win/external/wpt/html/semantics/forms/the-select-element/select-value-expected.txt

This is a testharness.js-based test.
PASS options 
PASS optgroups 
PASS option is child of div 
PASS no options 
Harness: the test ran to completion.
